+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix for occasional failure to update `.meta` files into Warp kernel cache on Windows.
 - Fix the OpenGL renderer not being able to run without a CUDA device available
   ([GH-344](https://github.com/NVIDIA/warp/issues/344)).
+- Fix incorrect CUDA driver function versions ([GH-402](https://github.com/NVIDIA/warp/issues/402)).
 
 ## [1.5.0] - 2024-12-02